+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Fix check for coincident points (previously < ≈95mm got treated as coincident)
 
+### Added
+
+- LatLonNvectorSpherical.centreOf()
+
 ## [2.3.0] - 2021-11-16
 
 ### Fixed